+++ packages/core/src/components/regionalization/RegionalizationModal/index.ts
@@ -1,1 +1,2 @@
 export { default } from './RegionalizationModal'
+export { default as RegionalizationModalContent } from './RegionalizationModalContent'